+++ geotrek/tourism/filters.py
@@ -3,6 +3,7 @@ from django.utils.translation import ugettext_lazy as _
 import django_filters
 
 from geotrek.common.filters import StructureRelatedFilterSet
+from django.utils.datetime_safe import datetime
 
 from .models import TouristicContent, TouristicEvent
 
@@ -24,9 +25,32 @@ class BeforeFilter(django_filters.DateFilter):
         return qs.filter(begin_date__lte=value)
 
 
+class CompletedFilter(django_filters.BooleanFilter):
+    """
+    Filter events with end_date in past (event completed)
+    """
+    @property
+    def field(self):
+        field = super(CompletedFilter, self).field
+        field.initial = True
+        return field
+
+    def filter(self, qs, value):
+        queryset = qs
+
+        if value is True:
+            queryset = queryset.filter(end_date__lt=datetime.today)
+
+        elif value is False:
+            queryset = queryset.exclude(end_date__lt=datetime.today)
+
+        return queryset
+
+
 class TouristicEventFilterSet(StructureRelatedFilterSet):
     after = AfterFilter(label=_(u"After"))
     before = BeforeFilter(label=_(u"Before"))
+    completed = CompletedFilter(label=_(u"Completed"))
 
     class Meta(StructureRelatedFilterSet.Meta):
         model = TouristicEvent