+++ geotrek/tourism/tests/test_sync.py
@@ -22,13 +22,13 @@ def factory(factory, source):
 
 class SyncTest(TranslationResetMixin, TestCase):
     def test_sync(self):
-        source_a = RecordSourceFactory(name='A')
-        source_b = RecordSourceFactory(name='B')
+        source_a = RecordSourceFactory(name='Source A')
+        source_b = RecordSourceFactory(name='Source B')
         factory(TouristicContentFactory, source_a)
         factory(TouristicContentFactory, source_b)
         factory(TouristicEventFactory, source_a)
         factory(TouristicEventFactory, source_b)
-        management.call_command('sync_rando', settings.SYNC_RANDO_ROOT, url='http://localhost:8000', source='A', skip_tiles=True, verbosity='0')
+        management.call_command('sync_rando', settings.SYNC_RANDO_ROOT, url='http://localhost:8000', source='Source A', skip_tiles=True, verbosity='0')
 
         with open(os.path.join(settings.SYNC_RANDO_ROOT, 'api', 'en', 'touristiccontents.geojson'), 'r') as f:
             tcontents = json.load(f)