+++ geotrek/trekking/tests/test_views.py
@@ -1118,12 +1118,20 @@ class SyncRandoViewTest(TestCase):
         self.user = User.objects.create_user('bart', password='mahaha')
 
     def test_return_redirect(self):
-        response = self.client.get(reverse('trekking:sync_randos'))
+        response = self.client.get(reverse('trekking:sync_randos_view'))
         self.assertEqual(response.status_code, 302)
 
     def test_return_redirect_superuser(self):
         self.user.is_superuser = True
-        response = self.client.get(reverse('trekking:sync_randos'))
+        response = self.client.get(reverse('trekking:sync_randos_view'))
+        self.assertEqual(response.status_code, 302)
+
+    def test_post_sync_redirect(self):
+        """
+        test if sync can be launched by superuser post
+        """
+        self.user.is_superuser = True
+        response = self.client.post(reverse('trekking:sync_randos'))
         self.assertEqual(response.status_code, 302)