+++ geotrek/trekking/urls.py
@@ -11,7 +11,8 @@ from .views import (
     TrekDocumentPublic, POIDocumentPublic,
     TrekGPXDetail, TrekKMLDetail, WebLinkCreatePopup,
     CirkwiTrekView, CirkwiPOIView, TrekPOIViewSet,
-    SyncRandoRedirect, TrekServiceViewSet,
+    SyncRandoRedirect, TrekServiceViewSet, sync_view,
+    sync_update_json
 )
 from . import serializers as trekking_serializers
 
@@ -25,7 +26,9 @@ urlpatterns = patterns(
     url(r'^popup/add/weblink/', WebLinkCreatePopup.as_view(), name='weblink_add'),
     url(r'^api/cirkwi/circuits.xml', CirkwiTrekView.as_view()),
     url(r'^api/cirkwi/pois.xml', CirkwiPOIView.as_view()),
-    url(r'^commands/sync$', SyncRandoRedirect.as_view(), name='sync_randos')
+    url(r'^commands/sync$', SyncRandoRedirect.as_view(), name='sync_randos'),
+    url(r'^commands/syncview$', sync_view, name='sync_randos_view'),
+    url(r'^commands/statesync/$', sync_update_json, name='sync_randos_state'),
 )